requirements.txt
streamlit>=1.31
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
pillow>=10.0
matplotlib>=3.7
scikit-learn>=1.3
scipy>=1.11
opencv-python-headless>=4.8
